PatternFormatter specific-length source and optimizations.
Optimize the pattern formatter by doing time-zone calculations
once and using it's offset. Parse the format string once and
setup a vector of the fields wanted so parsing the %[name] is
much faster.

Add ability to easily enforce a specific length for the source
fields which is nicer for reading log files.