Adds support for building from paths with spaces.
Attempting to build POCO on Windows when it was sitting in a path with spaces (e.g., C:\my source\poco-1.4.6p2\) failed.  This would occur if you did the following from a VS command prompt:
cd "C:\my source\poco-1.4.6p2"
buildwin.cmd 110

The root cause was storing the output of the CD command in a local variable and then attempting to cd into that variable later without appropriately quoting it.