Added Timestamp::Min and Timestamp::Max
Two constants to initialize timestamps with some extreme values. The min
can be used if we are looking for the maximum timestamp from a set.
Notice, the *.vcxproj file changed since we need to add NOMINMAX to the
preprocessor on windows. Otherwise, lower case "min" and "max" are
macros from some Windows header and interfere with the
std::numerical_limits<>::min/max.