ReuseAddress and reusePort options should be separeted.
The reuseAddress option is to reuse a socket in TIME-WAIT state, bind still fails
if the socket is in-use, server applications often use the option.

But reusePort can reuse a port which is already in-use, I don't think you want
to share a port with another irrelevant application, unless you know what you
are doing, this is a more error prone option only is used in rare cases.