Fix longpathTest on Windows by add the NTFS prefix \\?\
NTFS filesystem supports paths up to 32k characters. You can use the win32 api and "\\?\" prefix the path to use greater than 260 characters.

A detailed explanation of long path from the .Net BCL team blog.
A small excerpt highlights the issue with long paths

Another concern is inconsistent behavior that would result by exposing long
path support. Long paths with the \\?\ prefix can be used in most of the file-related
Windows APIs, but not all Windows APIs. For example, LoadLibrary, which maps a module
into the address of the calling process, fails if the file name is longer than MAX_PATH.
So this means MoveFile will let you move a DLL to a location such that its path is
longer than 260 characters, but when you try to load the DLL, it would fail.
There are similar examples throughout the Windows APIs;
some workarounds exist, but they are on a case-by-case basis.


Signed-off-by: FrancisANDRE <zosrothko@orange.fr>