Fix for issue #1043
The fix for this issue was partially implemented in development branch.
However, there are still some issue, that were not covered:
1. The std::string overload will sets SQL_LONGVARCHAR as fSqlType if the
output parameter is of type VARCHAR(MAX)
2. The UTF16String overload of bind method:
a) always resolves fSqlType using the size of formal parameter const
std::string& val.
b) allocates half of the size of the SP's output parameter.
3. The _utf16Strings member is not sychronized in Binder::synchronize()
method.
4. The _utf16Strings member is not reset in Binder::reset() method.