Option to force PollingDirectoryWatcherStrategy
Add an option to force the use of PollingDirectoryWatcherStrategy
Should work on any platform
(Not Linux only)

See previous PR on this topic :
https://github.com/pocoproject/poco/pull/2881

On Linux, inotfy does not work for network volumes (such as NFS).

See https://stackoverflow.com/questions/4231243/inotify-with-nfs

By adding flag POCO_DW_FORCE_POLLING in Foundation/Config.h,
the use of PollingDirectoryWatcherStrategy is forced.

This is not the same behavior as flag POCO_NO_INOTIFY.
This only disables compilation of DirectoryWatcher.