allow to handle hot-journal (fixes #3135)
Before, when executing the Poco::Data::SQLite::Utility::fileToMemory method
for a SQLite database with an existing hot journal, the method returns successfully,
but the resulting in-memory database is corrupted.

This is because if a hot journal exists, this can not processed
correctly if only read only access to the database is available.