Fixed compilation error with MongoDB::Connection and Util::Application
MongoDB/Connection.h had to be included before Util/Application.h
on Windows with VS-2019, otherwise the compilation failed.
The root cause of this was that including Util/Application.h
introduced a Message type under the Poco::Util namespace, and
the MongoDB::MessageHeader class was friended with that type,
instead of MongoDB::Message. A simple solution for this is to
explicitly forward declare the Message class in the MongoDB namespace.