#3507: Fix bound configuration reference counting in Poco::Util::Option.
Type of _pConfig was changed from raw pointer to AutoPtr in commit 1bf40a0, but reference counting was not properly adapted. In most instances, refcount was incremented by 2, but not in "binding()". In the destructor, refcount was always decremented twice, leading to early deletion of the bound config and later possibly an invalid pointer access.