CppParser: Support for return values in global ns
If a return value of a virtual member function is specified to be in the
global namespace using leading double colons (e.g. virtual ::MyClass*)
the parser is now handling the virtual keyword correctly. Prior the
function was not declared as virtual and the return value was named as
virtual::MyClass* which led to completely missleading results.